refactor private_table_per_service@order_service
