+++ database_per_services/private_table_per_service/order_service/src/main/java/nawaphon/microservices/order_service/OrderServiceApplication.java
@@ -2,10 +2,8 @@ package nawaphon.microservices.order_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
 
 @SpringBootApplication
-@EntityScan("nawaphon.microservice.main.common.pojo")
 public class OrderServiceApplication {
 
     public static void main(String[] args) {
