+++ database_per_services/private_table_per_service/order_service/src/main/java/nawaphon/microservices/order_service/controllers/MainController.java
@@ -1,7 +1,7 @@
 package nawaphon.microservices.order_service.controllers;
 
-import nawaphon.microservices.order_service.controllers.pojo.Order;
-import nawaphon.microservices.order_service.controllers.pojo.ResponseMessage;
+import nawaphon.microservices.order_service.pojo.Order;
+import nawaphon.microservices.order_service.pojo.ResponseMessage;
 import nawaphon.microservices.order_service.services.MainService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
