+++ database_per_services/private_table_per_service/order_service/src/main/java/nawaphon/microservices/order_service/repositories/OrderRepository.java
@@ -1,6 +1,6 @@
 package nawaphon.microservices.order_service.repositories;
 
-import nawaphon.microservices.order_service.controllers.pojo.Order;
+import nawaphon.microservices.order_service.pojo.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
