+++ database_per_services/private_table_per_service/order_service/src/main/java/nawaphon/microservices/order_service/services/MainService.java
@@ -1,25 +1,12 @@
 package nawaphon.microservices.order_service.services;
 
-import nawaphon.microservices.order_service.controllers.pojo.Order;
-import nawaphon.microservices.order_service.controllers.pojo.ResponseMessage;
-import nawaphon.microservices.order_service.repositories.OrderRepository;
-import org.springframework.stereotype.Service;
+
+import nawaphon.microservices.order_service.pojo.Order;
+import nawaphon.microservices.order_service.pojo.ResponseMessage;
 
 import java.util.List;
 
-@Service
-public class MainService {
+public interface MainService {
 
-    private final OrderRepository orderRepository;
-
-    public MainService(final OrderRepository orderRepository) {
-        this.orderRepository = orderRepository;
-    }
-
-
-    public ResponseMessage<List<Order>> firstService() {
-        final List<Order> results = this.orderRepository.findAll();
-
-        return new ResponseMessage<>(200, "Done", results);
-    }
+    ResponseMessage<List<Order>> firstService();
 }
