--- database_per_services/private_table_per_service/order_service/src/test/java/nawaphon/microservices/order_service/PrivateTablePerServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package nawaphon.microservices.order_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PrivateTablePerServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
